requirements.txt
streamlit==1.37.1
youtube-transcript-api==0.6.2
groq==0.9.0
python-dotenv==1.0.1